requirements.txt
semver==2.13.0
docker==5.0.0
requests==2.25.1
python-dateutil==2.8.1
